[scanner] refactor: split MCP events.ts into focused modules
Split events.ts (574 lines) into focused domain-aligned modules:
- events/shared.ts (163 lines) - shared state, cache, demo data
- events/useEvents.ts (235 lines) - main Kubernetes events hook
- events/useWarningEvents.ts (170 lines) - warning events filtering
- events/index.ts - barrel re-export

Converted events.ts to 1-line barrel for backward compatibility.
Issues #18908, #18910, #18911 were already completed in prior commits.
