+++ web/src/hooks/mcp/events.ts
@@ -1,574 +1,1 @@
-import { useState, useEffect, useCallback, useRef } from 'react'
-import { reportAgentDataSuccess, isAgentUnavailable } from '../useLocalAgent'
-import { fetchSSE } from '../../lib/sseClient'
-import { isDemoMode } from '../../lib/demoMode'
-import { useDemoMode } from '../useDemoMode'
-import { registerRefetch } from '../../lib/modeTransition'
-import { registerCacheReset } from '../../lib/modeTransition'
-import { REFRESH_INTERVAL_MS, MIN_REFRESH_INDICATOR_MS, getEffectiveInterval, getLocalAgentURL, agentFetch } from './shared'
-import { subscribePolling } from './pollingManager'
-import { MCP_HOOK_TIMEOUT_MS, LOCAL_AGENT_HTTP_URL } from '../../lib/constants/network'
-import { CONSECUTIVE_FAILURE_THRESHOLD } from '../../lib/cache'
-import { isClusterModeBackend } from '../../lib/cache/fetcherUtils'
-import type { ClusterEvent } from './types'
-
-// ---------------------------------------------------------------------------
-// Shared Events State - enables cache reset notifications to all consumers
-// ---------------------------------------------------------------------------
-
-interface EventsSharedState {
-  cacheVersion: number
-  isResetting: boolean
-}
-
-let eventsSharedState: EventsSharedState = {
-  cacheVersion: 0,
-  isResetting: false,
-}
-
-type EventsSubscriber = (state: EventsSharedState) => void
-const eventsSubscribers = new Set<EventsSubscriber>()
-
-function notifyEventsSubscribers() {
-  Array.from(eventsSubscribers).forEach(subscriber => subscriber(eventsSharedState))
-}
-
-export function subscribeEventsCache(callback: EventsSubscriber): () => void {
-  eventsSubscribers.add(callback)
-  return () => eventsSubscribers.delete(callback)
-}
-
-// Module-level cache for events data (persists across navigation)
-interface EventsCache {
-  data: ClusterEvent[]
-  timestamp: Date
-  key: string
-}
-let eventsCache: EventsCache | null = null
-
-export function useEvents(cluster?: string, namespace?: string, limit = 20) {
-  const cacheKey = `events:${cluster || 'all'}:${namespace || 'all'}:${limit}`
-  const abortControllerRef = useRef<AbortController | null>(null)
-  const refreshTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const isMountedRef = useRef(true)
-  const { isDemoMode: demoMode } = useDemoMode()
-  const initialMountRef = useRef(true)
-
-  // Initialize from cache if available
-  const getCachedData = () => {
-    if (eventsCache && eventsCache.key === cacheKey) {
-      return { data: eventsCache.data, timestamp: eventsCache.timestamp }
-    }
-    return null
-  }
-
-  const cached = getCachedData()
-  const [events, setEvents] = useState<ClusterEvent[]>(cached?.data || [])
-  const [isLoading, setIsLoading] = useState(!cached)
-  const [isRefreshing, setIsRefreshing] = useState(false)
-  const [lastUpdated, setLastUpdated] = useState<Date | null>(cached?.timestamp || null)
-  const [error, setError] = useState<string | null>(null)
-  const [consecutiveFailures, setConsecutiveFailures] = useState(0)
-  const [lastRefresh, setLastRefresh] = useState<Date | null>(cached?.timestamp || null)
-
-  const clearRefreshTimeout = useCallback(() => {
-    if (refreshTimeoutRef.current) {
-      clearTimeout(refreshTimeoutRef.current)
-      refreshTimeoutRef.current = null
-    }
-  }, [])
-
-  const finishRefreshing = useCallback((silent: boolean) => {
-    clearRefreshTimeout()
-    if (silent) {
-      setIsRefreshing(false)
-      return
-    }
-    refreshTimeoutRef.current = setTimeout(() => {
-      refreshTimeoutRef.current = null
-      if (isMountedRef.current) {
-        setIsRefreshing(false)
-      }
-    }, MIN_REFRESH_INDICATOR_MS)
-  }, [clearRefreshTimeout])
-
-  const refetch = useCallback(async (silent = false) => {
-    // In demo mode, use demo data
-    if (isDemoMode()) {
-      const demoEvents = getDemoEvents().filter(e =>
-        (!cluster || e.cluster === cluster) && (!namespace || e.namespace === namespace)
-      ).slice(0, limit)
-      setEvents(demoEvents)
-      const now = new Date()
-      setLastUpdated(now)
-      setLastRefresh(now)
-      setIsLoading(false)
-      setError(null)
-      if (!silent) {
-        setIsRefreshing(true)
-      }
-      finishRefreshing(silent)
-      return
-    }
-
-    // For silent (background) refreshes, don't update loading states - prevents UI flashing
-    if (!silent) {
-      // Always set isRefreshing first so indicator shows
-      setIsRefreshing(true)
-      const hasCachedData = eventsCache && eventsCache.key === cacheKey
-      if (!hasCachedData) {
-        setIsLoading(true)
-      }
-    }
-
-    // Abort any previous in-flight request
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort()
-    }
-    abortControllerRef.current = new AbortController()
-    const signal = abortControllerRef.current.signal
-    const clusterModeBackend = isClusterModeBackend()
-
-    // Try direct REST endpoint first for single-cluster queries.
-    if (cluster && (clusterModeBackend || !isAgentUnavailable())) {
-      try {
-        const params = new URLSearchParams()
-        params.append('cluster', cluster)
-        if (namespace) params.append('namespace', namespace)
-        params.append('limit', limit.toString())
-
-        const timeoutId = setTimeout(() => abortControllerRef.current?.abort(), MCP_HOOK_TIMEOUT_MS)
-        const response = clusterModeBackend
-          ? await fetch(`/api/mcp/events?${params}`, {
-              signal: AbortSignal.timeout(MCP_HOOK_TIMEOUT_MS),
-              headers: { 'Accept': 'application/json' },
-            })
-          : await agentFetch(`${getLocalAgentURL()}/events?${params}`, {
-              signal,
-              headers: { 'Accept': 'application/json' },
-            })
-        clearTimeout(timeoutId)
-
-        if (response.ok) {
-          const data = await response.json()
-          const eventData = data.events || []
-          const now = new Date()
-          eventsCache = { data: eventData, timestamp: now, key: cacheKey }
-          setEvents(eventData)
-          setError(null)
-          setLastUpdated(now)
-          setConsecutiveFailures(0)
-          setLastRefresh(now)
-          setIsLoading(false)
-          finishRefreshing(silent)
-          if (!clusterModeBackend) {
-            reportAgentDataSuccess()
-          }
-          return
-        }
-      } catch (err: unknown) {
-        console.error(`[useEvents] ${clusterModeBackend ? 'Backend' : 'Local agent'} failed for ${cluster}:`, err)
-      }
-    }
-
-    // Use SSE streaming for progressive multi-cluster data
-    try {
-      const sseParams: Record<string, string> = {}
-      if (cluster) sseParams.cluster = cluster
-      if (namespace) sseParams.namespace = namespace
-      sseParams.limit = limit.toString()
-
-      const allEvents = await fetchSSE<ClusterEvent>({
-        url: `${clusterModeBackend ? '/api/mcp' : LOCAL_AGENT_HTTP_URL}/events/stream`,
-        params: sseParams,
-        itemsKey: 'events',
-        signal,
-        onClusterData: (_clusterName, items) => {
-          if (signal.aborted || !isMountedRef.current) return
-          setEvents(prev => [...prev, ...items].slice(0, limit))
-          setIsLoading(false)
-        },
-      })
-
-      if (!isMountedRef.current) return
-      const now = new Date()
-      eventsCache = { data: allEvents.slice(0, limit), timestamp: now, key: cacheKey }
-      setEvents(allEvents.slice(0, limit))
-      setError(null)
-      setLastUpdated(now)
-      setConsecutiveFailures(0)
-      setLastRefresh(now)
-    } catch (err: unknown) {
-      // Use name check instead of instanceof to handle both Error and DOMException
-      // across all browser versions (DOMException may not extend Error in older Safari).
-      if ((err as { name?: string })?.name === 'AbortError') return
-      if (!isMountedRef.current) return
-      setConsecutiveFailures(prev => prev + 1)
-      setLastRefresh(new Date())
-      if (!silent && !eventsCache) {
-        setError('Failed to fetch events')
-      }
-    } finally {
-      if (isMountedRef.current) {
-        setIsLoading(false)
-        finishRefreshing(silent)
-      }
-    }
-  }, [cacheKey, cluster, finishRefreshing, limit, namespace])
-
-  // Track mounted state for cleanup
-  useEffect(() => {
-    isMountedRef.current = true
-    return () => {
-      isMountedRef.current = false
-      clearRefreshTimeout()
-      if (abortControllerRef.current) {
-        abortControllerRef.current.abort()
-      }
-    }
-  }, [clearRefreshTimeout])
-
-  useEffect(() => {
-    const hasCachedData = eventsCache && eventsCache.key === cacheKey
-    refetch(!!hasCachedData)
-    return registerRefetch(`events:${cacheKey}`, () => refetch(false))
-  }, [cacheKey, refetch])
-
-  useEffect(() => {
-    return subscribePolling(
-      `events:${cacheKey}`,
-      getEffectiveInterval(REFRESH_INTERVAL_MS, consecutiveFailures),
-      () => refetch(true),
-    )
-  }, [cacheKey, consecutiveFailures, refetch])
-
-  // Subscribe to cache reset notifications - triggers skeleton when cache is cleared
-  useEffect(() => {
-    const handleCacheReset = (state: EventsSharedState) => {
-      if (state.isResetting) {
-        setIsLoading(true)
-        setEvents([])
-        setLastUpdated(null)
-      }
-    }
-    return subscribeEventsCache(handleCacheReset)
-  }, [])
-
-  // Re-fetch when demo mode changes (not on initial mount)
-  useEffect(() => {
-    if (initialMountRef.current) {
-      initialMountRef.current = false
-      return
-    }
-    refetch(false)
-  }, [demoMode, refetch])
-
-  return {
-    events,
-    isLoading,
-    isRefreshing,
-    lastUpdated,
-    error,
-    refetch: () => refetch(false),
-    consecutiveFailures,
-    isFailed: consecutiveFailures >= CONSECUTIVE_FAILURE_THRESHOLD,
-    lastRefresh,
-  }
-}
-
-// Module-level cache for warning events data (persists across navigation)
-interface WarningEventsCache {
-  data: ClusterEvent[]
-  timestamp: Date
-  key: string
-}
-let warningEventsCache: WarningEventsCache | null = null
-
-export function useWarningEvents(cluster?: string, namespace?: string, limit = 20) {
-  const cacheKey = `warningEvents:${cluster || 'all'}:${namespace || 'all'}:${limit}`
-  const abortControllerRef = useRef<AbortController | null>(null)
-  const refreshTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const isMountedRef = useRef(true)
-  const { isDemoMode: demoMode } = useDemoMode()
-  const initialMountRef = useRef(true)
-
-  // Initialize from cache if available
-  const getCachedData = () => {
-    if (warningEventsCache && warningEventsCache.key === cacheKey) {
-      return { data: warningEventsCache.data, timestamp: warningEventsCache.timestamp }
-    }
-    return null
-  }
-
-  const cached = getCachedData()
-  const [events, setEvents] = useState<ClusterEvent[]>(cached?.data || [])
-  const [isLoading, setIsLoading] = useState(!cached)
-  const [isRefreshing, setIsRefreshing] = useState(false)
-  const [lastUpdated, setLastUpdated] = useState<Date | null>(cached?.timestamp || null)
-  const [error, setError] = useState<string | null>(null)
-  const [consecutiveFailures, setConsecutiveFailures] = useState(0)
-
-  const clearRefreshTimeout = useCallback(() => {
-    if (refreshTimeoutRef.current) {
-      clearTimeout(refreshTimeoutRef.current)
-      refreshTimeoutRef.current = null
-    }
-  }, [])
-
-  const finishRefreshing = useCallback((silent: boolean) => {
-    clearRefreshTimeout()
-    if (silent) {
-      setIsRefreshing(false)
-      return
-    }
-    refreshTimeoutRef.current = setTimeout(() => {
-      refreshTimeoutRef.current = null
-      if (isMountedRef.current) {
-        setIsRefreshing(false)
-      }
-    }, MIN_REFRESH_INDICATOR_MS)
-  }, [clearRefreshTimeout])
-
-  const refetch = useCallback(async (silent = false) => {
-    // For silent (background) refreshes, don't update loading states - prevents UI flashing
-    if (!silent) {
-      // Always set isRefreshing first so indicator shows
-      setIsRefreshing(true)
-      const hasCachedData = warningEventsCache && warningEventsCache.key === cacheKey
-      if (!hasCachedData) {
-        setIsLoading(true)
-      }
-    }
-    // In demo mode, use demo data
-    if (isDemoMode()) {
-      const demoWarnings = getDemoEvents().filter(e =>
-        e.type === 'Warning' &&
-        (!cluster || e.cluster === cluster) &&
-        (!namespace || e.namespace === namespace)
-      ).slice(0, limit)
-      setEvents(demoWarnings)
-      const now = new Date()
-      setLastUpdated(now)
-      setError(null)
-      setIsLoading(false)
-      finishRefreshing(silent)
-      return
-    }
-
-    // Abort any previous in-flight request
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort()
-    }
-    abortControllerRef.current = new AbortController()
-    const signal = abortControllerRef.current.signal
-
-    // Use SSE streaming for progressive multi-cluster data
-    try {
-      const sseParams: Record<string, string> = {}
-      if (cluster) sseParams.cluster = cluster
-      if (namespace) sseParams.namespace = namespace
-      sseParams.limit = limit.toString()
-
-      // events/warnings is a backend-only endpoint (#9996) — route SSE via /api/mcp/
-      const allEvents = await fetchSSE<ClusterEvent>({
-        url: `/api/mcp/events/warnings/stream`,
-        params: sseParams,
-        itemsKey: 'events',
-        signal,
-        onClusterData: (_clusterName, items) => {
-          if (signal.aborted || !isMountedRef.current) return
-          setEvents(prev => [...prev, ...items].slice(0, limit))
-          setIsLoading(false)
-        },
-      })
-
-      if (signal.aborted || !isMountedRef.current) return
-      const now = new Date()
-      warningEventsCache = { data: allEvents.slice(0, limit), timestamp: now, key: cacheKey }
-      setEvents(allEvents.slice(0, limit))
-      setError(null)
-      setLastUpdated(now)
-      setConsecutiveFailures(0)
-    } catch (err: unknown) {
-      // Use name check instead of instanceof to handle both Error and DOMException
-      // across all browser versions (DOMException may not extend Error in older Safari).
-      if ((err as { name?: string })?.name === 'AbortError') return
-      if (!isMountedRef.current) return
-      setConsecutiveFailures(prev => prev + 1)
-      if (!silent && !warningEventsCache) {
-        setError('Failed to fetch warning events')
-      }
-    } finally {
-      if (isMountedRef.current && !signal.aborted) {
-        setIsLoading(false)
-        finishRefreshing(silent)
-      }
-    }
-  }, [cacheKey, cluster, finishRefreshing, limit, namespace])
-
-  // Track mounted state for cleanup
-  useEffect(() => {
-    isMountedRef.current = true
-    return () => {
-      isMountedRef.current = false
-      clearRefreshTimeout()
-      if (abortControllerRef.current) {
-        abortControllerRef.current.abort()
-      }
-    }
-  }, [clearRefreshTimeout])
-
-  useEffect(() => {
-    const hasCachedData = warningEventsCache && warningEventsCache.key === cacheKey
-    refetch(!!hasCachedData)
-    return registerRefetch(`warning-events:${cacheKey}`, () => refetch(false))
-  }, [cacheKey, refetch])
-
-  useEffect(() => {
-    return subscribePolling(
-      `warningEvents:${cacheKey}`,
-      getEffectiveInterval(REFRESH_INTERVAL_MS, consecutiveFailures),
-      () => refetch(true),
-    )
-  }, [cacheKey, consecutiveFailures, refetch])
-
-  // Subscribe to cache reset notifications - triggers skeleton when cache is cleared
-  useEffect(() => {
-    const handleCacheReset = (state: EventsSharedState) => {
-      if (state.isResetting) {
-        setIsLoading(true)
-        setEvents([])
-        setLastUpdated(null)
-      }
-    }
-    return subscribeEventsCache(handleCacheReset)
-  }, [])
-
-  // Re-fetch when demo mode changes (not on initial mount)
-  useEffect(() => {
-    if (initialMountRef.current) {
-      initialMountRef.current = false
-      return
-    }
-    refetch(false)
-  }, [demoMode, refetch])
-
-  return { events, isLoading, isRefreshing, lastUpdated, error, refetch: () => refetch(false) }
-}
-
-function getDemoEvents(): ClusterEvent[] {
-  const now = new Date()
-  const minutesAgo = (m: number) => new Date(now.getTime() - m * 60000).toISOString()
-
-  return [
-    {
-      type: 'Warning',
-      reason: 'FailedScheduling',
-      message: 'No nodes available to schedule pod',
-      object: 'Pod/worker-5c6d7e8f9-n3p2q',
-      namespace: 'batch',
-      cluster: 'vllm-gpu-cluster',
-      count: 3,
-      firstSeen: minutesAgo(25),
-      lastSeen: minutesAgo(5),
-    },
-    {
-      type: 'Normal',
-      reason: 'Scheduled',
-      message: 'Successfully assigned pod to node-2',
-      object: 'Pod/api-server-7d8f9c6b5-abc12',
-      namespace: 'production',
-      cluster: 'eks-prod-us-east-1',
-      count: 1,
-      firstSeen: minutesAgo(12),
-      lastSeen: minutesAgo(12),
-    },
-    {
-      type: 'Warning',
-      reason: 'BackOff',
-      message: 'Back-off restarting failed container',
-      object: 'Pod/api-server-7d8f9c6b5-x2k4m',
-      namespace: 'production',
-      cluster: 'eks-prod-us-east-1',
-      count: 15,
-      firstSeen: minutesAgo(45),
-      lastSeen: minutesAgo(2),
-    },
-    {
-      type: 'Normal',
-      reason: 'Pulled',
-      message: 'Container image pulled successfully',
-      object: 'Pod/frontend-8e9f0a1b2-def34',
-      namespace: 'web',
-      cluster: 'gke-staging',
-      count: 1,
-      firstSeen: minutesAgo(8),
-      lastSeen: minutesAgo(8),
-    },
-    {
-      type: 'Warning',
-      reason: 'Unhealthy',
-      message: 'Readiness probe failed: connection refused',
-      object: 'Pod/cache-redis-0',
-      namespace: 'data',
-      cluster: 'gke-staging',
-      count: 8,
-      firstSeen: minutesAgo(30),
-      lastSeen: minutesAgo(1),
-    },
-    {
-      type: 'Normal',
-      reason: 'ScalingReplicaSet',
-      message: 'Scaled up replica set api-gateway-7d8c to 3',
-      object: 'Deployment/api-gateway',
-      namespace: 'production',
-      cluster: 'eks-prod-us-east-1',
-      count: 1,
-      firstSeen: minutesAgo(18),
-      lastSeen: minutesAgo(18),
-    },
-    {
-      type: 'Normal',
-      reason: 'SuccessfulCreate',
-      message: 'Created pod: worker-5c6d7e8f9-abc12',
-      object: 'ReplicaSet/worker-5c6d7e8f9',
-      namespace: 'batch',
-      cluster: 'vllm-gpu-cluster',
-      count: 1,
-      firstSeen: minutesAgo(22),
-      lastSeen: minutesAgo(22),
-    },
-    {
-      type: 'Warning',
-      reason: 'FailedMount',
-      message: 'MountVolume.SetUp failed for volume "config": configmap "app-config" not found',
-      object: 'Pod/ml-inference-7f8g9h-xyz99',
-      namespace: 'ml',
-      cluster: 'vllm-gpu-cluster',
-      count: 4,
-      firstSeen: minutesAgo(35),
-      lastSeen: minutesAgo(3),
-    },
-  ]
-}
-
-// Register with mode transition coordinator for unified cache clearing
-if (typeof window !== 'undefined') {
-  registerCacheReset('events', () => {
-    // Set resetting flag to trigger skeleton display
-    eventsSharedState = {
-      cacheVersion: eventsSharedState.cacheVersion + 1,
-      isResetting: true,
-    }
-    notifyEventsSubscribers()
-
-    eventsCache = null
-    warningEventsCache = null
-
-    // Reset the resetting flag after a tick
-    setTimeout(() => {
-      eventsSharedState = { ...eventsSharedState, isResetting: false }
-      notifyEventsSubscribers()
-    }, 0)
-  })
-}
+export * from './events/index'
